// src/app/ITS/shared/interfaces/backend/locations.ts
export interface ILocations {
    "Asset_Location_ID": number;
    "Display_ID": string;
    "Creator": string;
    "Creation_Time": string;
    "Edit_Time": string;
    "Availability": number;
    "Status": string;
    "Name": string;

    /* IPAT additions */
    "Address": string;
    "Company_Name": string;
    "Company_ID": string;
    "Company_Address_ID": string;
    "Country": string;
    "PostCode": string;
}