// src/app/ITS/shared/interfaces/backend/Customer/CustomerAddress.ts
export interface ICustomerAddress {
    "CustomerAddress_ID": number;
    "Display_ID": string;
    "Creator": string;
    "Creation_Time": string;
    "Edit_Time": string;
    "Availability": number;
    "Status": string;
    "Type": string;
    "Country": string;
    "Address": string;
    "Postcode": string;
    "Contact_Person": string;
    "Contact_Person_HP": string;
    "Email": string;
    "Is_Depot": boolean;
    "Customer_ID": number;
    "Address_Name": string;
}